% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BarDiscrete.r
\name{BarDiscrete}
\alias{BarDiscrete}
\title{Create bar charts for discrete features}
\usage{
BarDiscrete(data, na.rm = TRUE, maxcat = 50)
}
\arguments{
\item{data}{input data to be plotted, in either \link{data.frame} or \link{data.table} format.}

\item{na.rm}{logical, indicating if missing values should be removed for each feature. The default is \code{TRUE}.}

\item{maxcat}{maximum categories allowed for each feature. The default is 50. More information in 'Details' section.}
}
\description{
This function creates frequency bar charts for each discrete feature.
}
\details{
If a discrete feature contains more categories than \code{maxcat} specifies, it will not be passed to the plotting function. Instead, it will be passed to \code{cat} with number of categories.
}
\examples{
# load packages
library(ggplot2)
library(data.table)
# load diamonds dataset from ggplot2
data("diamonds")

# plot bar charts for diamonds dataset
BarDiscrete(diamonds)

# making more columns as factors
diamonds <- data.table(diamonds)
diamonds[, color_clarity:=as.factor(paste0(color, "_", clarity))]
diamonds[, cut_color:=as.factor(paste0(cut, "_", color))]
diamonds[, cut_clarity:=as.factor(paste0(cut, "_", clarity))]
diamonds[, color_clarity:=as.factor(paste0(color, "_", clarity))]
diamonds2 <- dcast.data.table(diamonds,
                              carat+cut+color+clarity+cut_color+cut_clarity+color_clarity+depth+table+x+y+z~color_clarity,
                              fun=sum,
                              value.var="price",
                              fill=NA)
for (col in names(diamonds2)[grep("D_|E_|F_|G_|H_|I_|J_", names(diamonds2))]) {
  set(diamonds2, j=col, value=as.factor(diamonds2[[col]]))
}
# plot bar charts for diamonds2 dataset
BarDiscrete(diamonds2, maxcat=100)
}
\keyword{bardiscrete}

