% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BoxplotContinuous.r
\name{BoxplotContinuous}
\alias{BoxplotContinuous}
\title{Create boxplot for continuous features}
\usage{
BoxplotContinuous(data, five_points_only = FALSE)
}
\arguments{
\item{data}{input data to be plotted, in either \link{data.frame} or \link{data.table} format.}

\item{five_points_only}{logical, indicating if plot with only minimum, Q1, Q2, Q3 and maximum, i.e., 5 points. The default is \code{FALSE}. More information in 'Details' section.}
}
\description{
This function creates boxplot for each continuous feature.
}
\details{
Setting \code{five_points_only} to \code{TRUE} could improve performance dramatically when dealing with large skewed datasets. Outliers (except for minimum and maximum) and whiskers may not be plotted.
}
\examples{
# plot using iris data
BoxplotContinuous(iris)
BoxplotContinuous(iris, five_points_only=TRUE)

# plot using random data
data <- data.table(matrix(rnorm(16000000), ncol=16))
BoxplotContinuous(data)
BoxplotContinuous(data, five_points_only=TRUE)
}
\keyword{boxplotcontinuous}

