% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExplorer-package.r
\docType{package}
\name{DataExplorer-package}
\alias{DataExplorer}
\alias{DataExplorer-package}
\alias{dataexplorer}
\title{Data Explorer}
\description{
Simplify and automate EDA process and report generation.
}
\details{
Data exploration process for data analysis and model building, so
that users could focus on understanding data and extracting insights. The
package automatically scans through each variable and does data profiling.
Typical graphical techniques will be performed for both discrete and
continuous features.
}
\seealso{
\link{GenerateReport},
\link{PlotMissing},
\link{BarDiscrete},
\link{HistogramContinuous},
\link{PlotStr},
\link{CollapseCategory},
\link{SetNaTo}
}

