% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dexplr}
\alias{dexplr}
\title{Explore a data frame visually}
\usage{
dexplr(dataset)
}
\arguments{
\item{dataset}{a data frame to explore}
}
\description{
Explore my dataset visually using an interactive shiny app
}
\examples{
library(MASS)
dexplr(Aids2)
}
\keyword{5D}
\keyword{5d}
\keyword{exploratory}
\keyword{explore}
\keyword{five-dimensional}
\keyword{shiny}
\keyword{visualize}

