% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{.getPageLayout}
\alias{.getPageLayout}
\title{Calculate page layout index}
\usage{
.getPageLayout(nrow, ncol, n)
}
\arguments{
\item{nrow}{number of rows per page}

\item{ncol}{number of columns per page}

\item{n}{number of features}
}
\value{
a list containing column indices for each page
}
\description{
Calculate column index on each page based on row and column counts
}
