% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{.ignoreCat}
\alias{.ignoreCat}
\title{Truncate category}
\usage{
.ignoreCat(dt, maxcat)
}
\arguments{
\item{dt}{input data object.}

\item{maxcat}{maximum categories allowed for each discrete feature.}
}
\value{
a named vector containing indices of features to be ignored.
}
\description{
Output index and name for features that will be ignored
}
