% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{.lapply}
\alias{.lapply}
\title{Parallelization}
\usage{
.lapply(parallel, X, FUN)
}
\arguments{
\item{parallel}{enable parallel?}

\item{X}{a vector (atomic or list) or an expression object}

\item{FUN}{the function to be applied to each element of X}
}
\value{
list of output from input
}
\description{
Parallelize jobs based on condition
}
