\name{explr-package}
\alias{explr-package}
\alias{explr}
\docType{package}
\title{
I allow visual exploration of a given dataset
}
\description{
I try to interpret variables as categorical or continuous
Then allow one to visualize up to five at once, at most 2
continuous and 3 categorical.
}
\details{
\tabular{ll}{
Package: \tab explr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-07-22\cr
License: \tab MIT\cr
}
library(MASS) # load the aids dataset
explr(aids)
}
\author{
John Collins
Maintainer: John Collins <johnssocks@gmail.com>
}
\references{
}
\keyword{ package, shiny, explore, exploration }
\seealso{
}
\examples{
}
