% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filter.by.type}
\alias{filter.by.type}
\title{filter.by.type}
\usage{
filter.by.type(dataset, classes.names, names.list)
}
\arguments{
\item{dataset}{a data frame}

\item{classes.names}{vector of allowable column types}

\item{names.list}{list of vectors representing the classes of each column}
}
\description{
Utility function to return a subset of names from a data.frame
which match the recignized classes in classes.list, given that
each name may have multiple classes in names.list ( a list)
}
\examples{
X <- data.frame(X=factor(c(1,2,1,2)), Y=c(1.1, 1.2, pi, 23.*2))
filter.by.type(X, c('factor', 'ordinal'), list(X=c('factor'), Y=c('numeric')))
}
\keyword{filter}
\keyword{names}

