% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenerateReport.r
\name{GenerateReport}
\alias{GenerateReport}
\title{GenerateReport Function}
\usage{
GenerateReport(input_data, output_file = "report.html",
  output_dir = getwd(), ...)
}
\arguments{
\item{input_data}{data source to be profiled, in either \link{data.frame} or \link{data.table} format.}

\item{output_file}{output file name. The default is "report.html".}

\item{output_dir}{output directory for report. The default is user's current directory.}

\item{\dots}{other arguments to be passed to \link{render}.}
}
\description{
This function generates the report of data profiling.
}
\examples{
# load library
library(rmarkdown)
library(ggplot2)
library(data.table)

# generate report for diamonds dataset
data(diamonds)
GenerateReport(diamonds,
               output_file="report.html",
               output_dir=getwd(),
               html_document(toc=TRUE, theme="flatly"))

# generate report for
data <- read.csv("http://www.cftc.gov/dea/newcot/deacit.txt", header=FALSE)[, -c(2:3, 32)]
GenerateReport(data)
}
\keyword{generatereport}

