% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce.r
\name{introduce}
\alias{introduce}
\title{Describe basic information}
\usage{
introduce(data)
}
\arguments{
\item{data}{input data.}
}
\value{
Describe basic information in a \link{data.frame}.
}
\description{
Describe basic information for input data.
}
\examples{
introduce(mtcars)
}
\keyword{introduce}
