% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_file.R
\name{mod_upload_file_server}
\alias{mod_upload_file_server}
\title{mod_upload_file_server}
\usage{
mod_upload_file_server(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
mod_upload_file_server
}
