% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.DataExplorerGrid}
\alias{plot.DataExplorerGrid}
\title{Plot objects with gridExtra}
\usage{
\method{plot}{DataExplorerGrid}(obj_list, page_layout, nrow, ncol, title,
  ggtheme, theme_config)
}
\arguments{
\item{obj_list}{list of ggplot objects}

\item{page_layout}{output of \link{.getPageLayout}}

\item{nrow}{number of rows per page}

\item{ncol}{number of columns per page}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes}

\item{theme_config}{a list of configurations to be passed to \link{theme}}
}
\value{
invisibly return the named list of ggplot objects
}
\description{
Plot multiple DataExplorer objects with the defined layout
}
