% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intro.r
\name{plot_intro}
\alias{plot_intro}
\title{Plot introduction}
\usage{
plot_intro(data, title = NULL, ggtheme = theme_gray(),
  theme_config = list())
}
\arguments{
\item{data}{input data}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes. The default is \link{theme_gray}.}

\item{theme_config}{a list of configurations to be passed to \link{theme}.}
}
\value{
invisibly return the ggplot object
}
\description{
Plot basic information (from \link{introduce}) for input data.
}
\details{
To change default font family and size, you may pass \code{base_size} and \code{base_family} to \code{ggtheme} options, e.g., \code{ggtheme = theme_gray(base_size = 15, base_family = "serif")}

\code{theme_config} argument expects all inputs to be wrapped in a list object, e.g., to change the text color: \code{theme_config = list("text" = element_text(color = "blue"))}
}
\examples{
plot_intro(airquality)
plot_intro(iris)
}
\seealso{
\link{introduce}
}
\keyword{plot_intro}
