% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotDataExplorer.multiple}
\alias{plotDataExplorer.multiple}
\title{Plot multiple objects}
\usage{
\method{plotDataExplorer}{multiple}(obj_list, page_layout, title, ggtheme,
  theme_config, ...)
}
\arguments{
\item{obj_list}{list of ggplot objects}

\item{page_layout}{output of \link{.getPageLayout}}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes}

\item{theme_config}{a list of configurations to be passed to \link{theme}}

\item{\dots}{other arguments to be passed to \link{facet_wrap}.}
}
\value{
invisibly return the named list of ggplot objects
}
\description{
Plot multiple DataExplorer objects with the defined layout
}
