% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotDataExplorer}
\alias{plotDataExplorer}
\title{Default DataExplorer plotting function}
\usage{
plotDataExplorer(...)
}
\arguments{
\item{\dots}{arguments to be passed}
}
\description{
Plot various DataExplorer objects
}
