% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotDataExplorer.single}
\alias{plotDataExplorer.single}
\title{Plot single object}
\usage{
\method{plotDataExplorer}{single}(obj, title, ggtheme, theme_config)
}
\arguments{
\item{obj}{ggplot object}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes}

\item{theme_config}{a list of configurations to be passed to \link{theme}}
}
\value{
invisibly return the ggplot object
}
\description{
Plot single DataExplorer object
}
