% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotMissing.r
\name{PlotMissing}
\alias{PlotMissing}
\title{PlotMissing Function}
\usage{
PlotMissing(data)
}
\arguments{
\item{data}{data to be analyzed}
}
\description{
This function plots frequency of missing values for each feature.
}
\examples{
PlotMissing()
}
\keyword{plotmissing}

