% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStr.r
\name{PlotStr}
\alias{PlotStr}
\title{Visualize data structure}
\usage{
PlotStr(data, type = c("diagonal", "radial"), max_level, ...)
}
\arguments{
\item{data}{input data}

\item{type}{type of network diagram. Defaults to \link{diagonalNetwork}.}

\item{max_level}{integer threshold of nested level to be visualized. Minimum 1 nested level and defaults to all.}

\item{\dots}{other arguments to be passed to plotting functions. See \link{diagonalNetwork} and \link{radialNetwork}.}
}
\description{
Visualize data structures in D3 network graph
}
\examples{
## Visualize structure of iris dataset
PlotStr(iris)

## Generate complicated data object
obj <- list(
  "a" = list(iris, airquality, list(mtcars = mtcars, USArrests = USArrests)),
  "b" = list(list(ts(1:10, frequency = 4))),
  "c" = lm(rnorm(5) ~ seq(5)),
  "d" = lapply(1:5, function(x) return(as.function(function(y) y + 1)))
)
## Visualize data object with radial network
PlotStr(obj, type = "r")
## Visualize only top 2 nested levels
PlotStr(obj, type = "r", max_level = 2)
}
\keyword{plotstr}

