% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_missing.R
\name{profile_missing}
\alias{profile_missing}
\title{Profile missing values}
\usage{
profile_missing(data)
}
\arguments{
\item{data}{input data}
}
\value{
missing value profile, such as frequency, percentage and suggested action.
}
\description{
Analyze missing value profile
}
\examples{
profile_missing(airquality)
}
\seealso{
\link{plot_missing}
}
\keyword{profile_missing}
