% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationContinuous.r
\name{CorrelationContinuous}
\alias{CorrelationContinuous}
\title{Create correlation heatmap for continuous features}
\usage{
CorrelationContinuous(data, ...)
}
\arguments{
\item{data}{input data to be plotted, in either \link{data.frame} or \link{data.table} format.}

\item{\dots}{other arguments to be passed to \link{cor}.}
}
\description{
This function creates a correlation heatmap for all continuous features.
}
\examples{
# correlation of features from mtcars dataset
data(mtcars)
CorrelationContinuous(mtcars)
}
\keyword{correlationcontinuous}

