% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExplorer-package.r
\docType{package}
\name{DataExplorer-package}
\alias{DataExplorer}
\alias{DataExplorer-package}
\title{DataExplorer: Automate Data Exploration and Treatment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automated data exploration process for analytic tasks and predictive modeling, so that users could focus on understanding data and extracting insights. The package scans and analyzes each variable, and visualizes them with typical graphical techniques. Common data processing methods are also available to treat and format data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://boxuancui.github.io/DataExplorer/}
  \item Report bugs at \url{https://github.com/boxuancui/DataExplorer/issues}
}

}
\author{
\strong{Maintainer}: Boxuan Cui \email{boxuancui@gmail.com}

}
\keyword{internal}
