% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dexplr}
\alias{dexplr}
\title{Explore a data frame visually}
\usage{
dexplr(dataset, ignore.names = c(), categorical.names = c(),
  continuous.names = c())
}
\arguments{
\item{dataset}{a data frame to explore}

\item{ignore.names}{column names to ignore in exploration}

\item{categorical.names}{column names to treat as categorical (discrete)}

\item{continuous.names}{column names to treat as continuous}
}
\description{
Explore my dataset visually using an interactive shiny app
}
\examples{
\dontrun{
library(MASS)
dexplr(Aids2)
q(save='no')
}
}
\keyword{5D}
\keyword{5d}
\keyword{exploratory}
\keyword{explore}
\keyword{five-dimensional}
\keyword{shiny}
\keyword{visualize}

