% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_columns.r
\name{drop_columns}
\alias{drop_columns}
\alias{DropVar}
\title{Drop selected variables}
\usage{
drop_columns(data, ind)
}
\arguments{
\item{data}{input data}

\item{ind}{a vector of either names or column positions of the variables to be dropped.}
}
\description{
Quickly drop variables by either name or column position.
}
\details{
\bold{This function updates \link{data.table} object directly.} Otherwise, output data will be returned matching input object class.
}
\examples{
# load packages
library(data.table)

# generate data
dt <- data.table(sapply(setNames(letters, letters), function(x) {assign(x, rnorm(100))}))
dt2 <- copy(dt)

# drop variables by name
names(dt)
drop_columns(dt, letters[2:25])
names(dt)

# drop variables by column position
names(dt2)
drop_columns(dt2, seq(2, 25))
names(dt2)

# work with non-data.table objects
iris_df <- data.table(iris)
drop_columns(iris_df, "Species")
class(iris_df) <- "data.frame"
}
\keyword{drop_columns}
