\docType{package}
\name{eda-package}
\alias{eda}
\alias{EDA}
\alias{eda-package}

\title{R Exploratory Data Analysis}

\description{
Automate regular EDA process and report generation.
}

\details{
The \pkg{eda} package automates most of EDA (Exploratory Data Analysis)
process for data analysis and model building, so that users could focus on
understanding data and extracting insights.
}

\seealso{
\link{GenerateReport},
\link{PlotMissing},
\link{BarDiscrete},
\link{HistogramContinuous},
\link{CorrelationDiscrete},
\link{CorrelationContinuous}
}
