% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-package.r
\docType{package}
\name{eda-package}
\alias{EDA}
\alias{eda}
\alias{eda-package}
\title{Exploratory Data Analysis}
\description{
Simplify and automate EDA process and report generation.
}
\details{
EDA (Exploratory Data Analysis) process for data analysis and model
building, so that users could focus on understanding data and extracting
insights. The package automatically scans through each variable and does
data profiling. Typical graphical techniques will be performed for both
discrete and continuous features.
}
\seealso{
\link{PlotMissing},
\link{BarDiscrete},
\link{HistogramContinuous},
\link{DensityContinuous},
\link{GenerateReport}
}

