% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistogramContinuous.r
\name{HistogramContinuous}
\alias{HistogramContinuous}
\title{Create histogram for continuous features}
\usage{
HistogramContinuous(data)
}
\arguments{
\item{data}{input data to be plotted, in either \link{data.frame} or \link{data.table} format.}
}
\description{
This function creates histogram for each continuous feature.
}
\examples{
# load library
library(data.table)

# plot using iris data
HistogramContinuous(iris)

# plot using random data
data <- data.table(matrix(rnorm(160000), ncol=16))
HistogramContinuous(data)
}
\keyword{histogramcontinuous}

