% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intro.r
\name{plot_intro}
\alias{plot_intro}
\title{Plot introduction}
\usage{
plot_intro(data, title = NULL, ggtheme = theme_gray(),
  theme_config = list())
}
\arguments{
\item{data}{input data}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes. The default is \link{theme_gray}.}

\item{theme_config}{a list of configurations to be passed to \link{theme}.}
}
\value{
invisibly return the ggplot object
}
\description{
Plot basic information (from \link{introduce}) for input data.
}
\examples{
plot_intro(airquality)
plot_intro(iris)
}
\seealso{
\link{introduce}
}
\keyword{plot_intro}
