% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.r
\name{plot_missing}
\alias{plot_missing}
\alias{PlotMissing}
\title{Plot missing value profile}
\usage{
plot_missing(data, title = NULL)
}
\arguments{
\item{data}{input data to be profiled, in either \link{data.frame} or \link{data.table} format.}

\item{title}{plot title}
}
\value{
missing value information, such as frequency, percentage and suggested action.
}
\description{
This function returns and plots frequency of missing values for each feature.
}
\details{
The returned object is suppressed by \link{invisible}.
}
\examples{
# Load packages
library(data.table)

# Add missing values to iris data
dt <- data.table(iris)
for (j in 1:4) set(dt, i = sample(150, j * 30), j, value = NA_integer_)

# Plot and assign missing value information
na_profile <- plot_missing(dt)
na_profile

# Drop columns with more than 50\% missing values
drop_columns(dt, as.character(na_profile[pct_missing >= 0.5][["feature"]]))
plot_missing(dt)
}
\keyword{plot_missing}
