% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.r
\name{plot_missing}
\alias{plot_missing}
\title{Plot missing value profile}
\usage{
plot_missing(data, title = NULL, ggtheme = theme_gray(),
  theme_config = list(legend.position = c("bottom")))
}
\arguments{
\item{data}{input data}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes. The default is \link{theme_gray}.}

\item{theme_config}{a list of configurations to be passed to \link{theme}.}
}
\value{
invisibly return the ggplot object
}
\description{
This function returns and plots frequency of missing values for each feature.
}
\details{
The returned object is suppressed by \link{invisible}.

To change default font family and size, you may pass \code{base_size} and \code{base_family} to \code{ggtheme} options, e.g., \code{ggtheme = theme_gray(base_size = 15, base_family = "serif")}

\code{theme_config} argument expects all inputs to be wrapped in a list object, e.g., to change the text color: \code{theme_config = list("text" = element_text(color = "blue"))}
}
\examples{
plot_missing(airquality)

# Plot with preset ggplot2 themes
library(ggplot2)
plot_missing(airquality, ggtheme = theme_light())
plot_missing(airquality, ggtheme = theme_minimal(base_size = 20))

# Plot with customized theme components
plot_missing(airquality, theme_config = list("legend.position" = c("top")))
}
\seealso{
\link{profile_missing}
}
\keyword{plot_missing}
