% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotDataExplorer.grid}
\alias{plotDataExplorer.grid}
\title{Plot objects with gridExtra}
\usage{
\method{plotDataExplorer}{grid}(
  plot_obj,
  title,
  ggtheme,
  theme_config,
  page_layout,
  nrow,
  ncol,
  ...
)
}
\arguments{
\item{plot_obj}{list of ggplot objects}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes}

\item{theme_config}{a list of configurations to be passed to \link{theme}}

\item{page_layout}{a list of page indices with associated plot indices}

\item{nrow}{number of rows per page}

\item{ncol}{number of columns per page}

\item{\dots}{other arguments to be passed}
}
\value{
invisibly return the named list of ggplot objects
}
\description{
Plot multiple DataExplorer objects using grid.arrange
}
\seealso{
\link{plotDataExplorer} \link{plotDataExplorer.single} \link{plotDataExplorer.multiple}
}
\keyword{internal}
