% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDataExplorer.multiple}
\alias{plotDataExplorer.multiple}
\title{Plot multiple objects}
\usage{
\method{plotDataExplorer}{multiple}(
  plot_obj,
  title,
  ggtheme,
  theme_config,
  page_layout,
  facet_wrap_args = list(),
  ...
)
}
\arguments{
\item{plot_obj}{list of ggplot objects separated by page}

\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes}

\item{theme_config}{a list of configurations to be passed to \link[ggplot2]{theme}}

\item{page_layout}{a list of page indices with associated plot indices}

\item{facet_wrap_args}{a list of other arguments to \link[ggplot2]{facet_wrap}}

\item{\dots}{other arguments to be passed}
}
\value{
invisibly return the named list of ggplot objects
}
\description{
Plot multiple DataExplorer objects with the defined layout
}
\seealso{
\link{plotDataExplorer} \link{plotDataExplorer.grid} \link{plotDataExplorer.single}
}
\keyword{internal}
