% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotDataExplorer.single}
\alias{plotDataExplorer.single}
\title{Plot single object}
\usage{
\method{plotDataExplorer}{single}(plot_obj, title, ggtheme, theme_config,
  ...)
}
\arguments{
\item{title}{plot title}

\item{ggtheme}{complete ggplot2 themes}

\item{theme_config}{a list of configurations to be passed to \link{theme}}

\item{\dots}{other arguments to be passed}

\item{obj}{single ggplot object}
}
\value{
invisibly return the ggplot object
}
\description{
Plot single DataExplorer object
}
\keyword{internal}
