% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SplitColType.r
\name{SplitColType}
\alias{SplitColType}
\title{Split data into discrete and continuous}
\usage{
SplitColType(data)
}
\arguments{
\item{data}{input data to be split, in either \code{\link{data.frame}} or \code{\link{data.table}} format.}
}
\value{
\code{discrete} all discrete features in \code{\link{data.table}} format

\code{continous} all continuous features in \code{\link{data.table}} format

\code{num_discrete} number of discrete features

\code{num_continuous} number of continuous features
}
\description{
This function splits the input data into two \code{\link{data.table}} objects: discrete and continuous. A feature is continuous if \code{is.numeric()} returns \code{TRUE}.
}
\examples{
output <- SplitColType(iris)
output$discrete
output$continuous
output$num_discrete
output$num_continuous
}
\keyword{splitcoltype}

